requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
fastapi
uvicorn
pymongo
boto3
python-jose
passlib[bcrypt]
python-multipart
email-validator
pydantic[email]
neo4j
faiss-cpu
cassandra-driver
fastapi
uvicorn
python-multipart
boto3
pymongo
python-dotenv
passlib[bcrypt]
jose
